Add initializeProgress, update test and README.md
A single initialize function now creates the progressArray monitor, opens a remote channel, and runs the progress monitor asynchronously which has been renamed from updateProgress! to monitorProgress!.  The test suite now uses the initializer function and tests cases where the number of workers exceeds and is less than the number of tasks.