Add SharedArray monitor and ETA
Instead of using a remote channel to push progress updates, initializes a shared array and passes that to the serialTask functions.  Calculation of progress percent, management of delay, and ETA calculation now done in monitorProgress!.  serialTask inner loop updates now occur every iteration and only increment the shared array.  Test file renamed to emphasize that multiple tasks are running in parallel as opposed to a single task in a parallel loop.